linux: Reduce spamming during shell initialization
tbot wants to know when a shell is ready on the remote side.  To do
this, it periodically probes the channel with a sequence to which the
shell responds in a recognizable way.

We want to detect the shell as fast as possible so up to now, the
sequence was sent every 0.2 seconds.

When the remote does not respond that fast, though, this creates a lot
of spam and a lot of probes for the remote shell to work through.  We
can actually slow down the shell initialization this way which we really
don't want to do.

As a stop-gap, let's increase the time from 0.2 to 3 seconds after the
first attempt.  This should still allow detecting fast shells really
quickly while not spamming slow shells/machines too much.  The exact
values/timeout-curve might need further tweaking in the future...

Signed-off-by: Harald Seiler <hws@denx.de>