channel: Don't error when exiting closed channel
When the remote side closes a channel, tests will usually abort.  This
includes exiting the context of the channel where a local closing of the
channel is attempted.  This of course fails because the channel is
already closed.

Only attempt this closing of the channel on context exit when the
channel was not already closed.

Signed-off-by: Harald Seiler <hws@denx.de>