subprocess: Always detect closed channel during read
We must ensure that tbot always detects when the channel is closed. This
is especially important during reads where we are of course waiting for
the remote end to respond.

The proper way to check for a closed channel is to check whether the
subprocess is still alive, as is done in the `closed` imeplementation.
This check must be performed periodically (alternative would be
SIGCHLD).

To get this working, we need to limit the time spent in `select()` and
instead wake up periodically to poll the child process.  The "polling
period" must be selected such that tbot sleeps enough to not waste
resources but also polls fast enough to not wait for a dead subprocess
unnecessarily long.  For now, let's use 300ms as a polling period.

Signed-off-by: Harald Seiler <hws@denx.de>