board.uboot: Fix shell init when serial line is slow
When the serial line is very slow, we can get into a situation where
tbot has spammed ^C a lot and then sees the first prompt show up.  Once
it does, tbot will attempt sending commands but the other side is still
processing the ^Cs.  This processing will then clobber the output of the
first commands which confuses tbot.

Fix this by giving the other side some time to react to the ^Cs we send.

Signed-off-by: Harald Seiler <hws@denx.de>