linux: Move copy() implementation into machine.linux
The tc.shell.copy() function is a very integral part of tbot so it
should really be a part of the machine.linux module instead.

Move it there but keep the old function as a compatibility alias.  You
can use it in the new location like this:

	from tbot.machine import linux

	linux.copy(path_src, path_dest)

For the linux.copy() version, do not make it a "testcase" like it was
previously.  This is more in-line with the rest of the module.

Signed-off-by: Harald Seiler <hws@denx.de>