linux: Support multiplexing in copy()
When the copy() function performs an scp transfer, it current does not
enable ssh multiplexing even if the machine is set up for it.  We can
speed up copy() here by doing this right.

When the relevant host has `use_multiplexing` set, also add the required
SSH options to the scp commandline to make the transfer use multiplexing
as well.

Signed-off-by: Harald Seiler <hws@denx.de>