log: Add only_decrease parameter to with_verbosity()
This parameter enables a mode where only lower verbosity levels are
actually applied.  This can be used in situations where a function never
wants to become _more_ verbose than the global level but wants to
_decrease_ verbosity for a certain amount of time.

A concrete example is disabling display of the command output for a few
commands.  Without with_verbosity(only_decrease=True), the code could
accidentally increase verbosity when the global level was QUIET.  The
only_decrease parameter prevents this.

Signed-off-by: Harald Seiler <hws@denx.de>