contrib: connector: auto: Add AutoConsoleConnector
The `AutoConsoleConnector` is a connector which will automatically
select the best available tool to open a serial console.  The tbot user
does not need to concern themselves with the specifics of each tool but
can instead define generic settings which will be passed correctly to
the chosen tool.

For now, the following tools are supported and selected in the order
below:

- tio: https://github.com/tio/tio
- picocom: https://github.com/npat-efault/picocom

Signed-off-by: Harald Seiler <hws@denx.de>