machine: u-boot: Add a hacky workaround for the crc32 command
The `crc32` command in U-Boot unfortunately prints the character
sequence `=> ` as part of its output.  This character sequence is an
often used U-Boot prompt, which means tbot's prompt-searching code gets
confused by `crc32` output.

Unfortunately, there is no real solution to this.  At least none that
works universally.  So to at least ease the pain a bit for people
running into this, attempt automatically applying a workaround for this
specific situation.

Signed-off-by: Rahix <rahix@rahix.de>