machine: linux: Make wait_for_shell() more robust
Slightly change the method used by wait_for_shell() to detect a response
from the remote shell.  Use a \ escape instead of '' quotation marks to
prevent a situation where only an odd number of quotation marks get
processed by the remote shell. Such a situation would lead to never
getting an appropriate response.

This issue was not properly confirmed to actually happen outside of
unreliable connections.  But for the sake of robustness, let's protect
ourselves against it anyway.

Reported-by: Suraj Krishnan <72937403+surajkrishnan14@users.noreply.github.com>
Signed-off-by: Rahix <rahix@rahix.de>