machine: linux: Add sanity check for shell interaction
Add a sanity check that ensures the shell on the remote end is really in
the state we expect it to be in before we start interacting with it.

This check should uncover the following kinds of errors early:

 - Heavy amounts of output clobbering from some source
 - tbot tracking the wrong prompt (offset error)
 - fancy shell configuration that tbot wasn't able to disable

Previously, such errors would only surface later, when trying to run
commands.  The error messages were quite confusing and didn't
immediately point to the root cause.  This sanity check should hopefully
make debugging these issues much easier.

Signed-off-by: Rahix <rahix@rahix.de>