selftest: Fix mock SSH server sometimes not being ready in time
Make sure the SSH server has some time to start up before trying to
connect to it.  We saw failures on the first test running against the
server, suggesting a startup race condition.

Ideally, this should be a proper availability check, but for now a small
delay will hopefully do.